requirements.txt
# Core Dependencies - Compatible versions
langchain==0.3.13
langchain-openai==0.2.14
langchain-community==0.3.13
langchain-core==0.3.28
langchain-text-splitters==0.3.4

# Vector Store - Updated to available version
faiss-cpu==1.12.0
# faiss-gpu==1.12.0  # Uncomment if you have GPU support

# OpenAI API
openai>=1.0.0

# Text Processing
tiktoken>=0.5.0

# Utilities
python-dotenv>=1.0.0
numpy>=1.24.0,<2.0.0
pydantic>=2.0.0,<3.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development (optional)
black>=23.0.0
flake8>=6.0.0
